requirements.txt
alembic==0.9.4
asn1crypto==0.22.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==2.0.2
enum34==1.1.6
Flask==0.12.2
Flask-Migrate==2.1.0
Flask-SQLAlchemy==2.2
future==0.16.0
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
MySQL-python==1.2.5
pkg-resources==0.0.0
pycparser==2.18
pyOpenSSL==17.2.0
python-dateutil==2.6.1
python-editor==1.0.3
python-telegram-bot==7.0.1
requests==2.18.2
six==1.10.0
SQLAlchemy==1.1.12
urllib3==1.22
Werkzeug==0.12.2
